コントローラーを分離
